Move the logic to a different class.
Move the logic to a different class to improve the readability.
